# backend/swagger.json
{
  "openapi": "3.0.1",
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Local server"
    },
    {
      "url": "https://api.pactsmith.com/api",
      "description": "Dev Server"
    },
    {
      "url": "https://api-prod.pactsmith.com/api",
      "description": "Prod Server"
    }
  ],
  "info": {
    "version": "0.0.1",
    "title": "nestjs-boilerplate",
    "description": "",
    "license": {
      "name": "UNLICENSED"
    }
  },
  "paths": {
...
    "/portals": {
      "post": {
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePortalDto"
              }
            }
          },
          "required": true,
          "x-nestia-encrypted": false
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Portals"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "x-nestia-namespace": "portals.createPortal",
        "x-nestia-jsDocTags": [],
        "x-nestia-method": "POST"
      },
      "get": {
        "tags": [],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "description": "",
            "required": true
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "description": "",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": ">}",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Readonly__type"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "summary": "Get all Portals",
        "description": "The code snippet you provided is a method in the `PortalsController` class. It is a route handler\nfor the GET request to `/portals` endpoint. Here's a breakdown of what it does:\nGets page",
        "x-nestia-namespace": "portals.getPortals",
        "x-nestia-jsDocTags": [
          {
            "name": "summary",
            "text": [
              {
                "text": "Get all Portals",
                "kind": "text"
              }
            ]
          },
          {
            "name": "date",
            "text": [
              {
                "text": "9/25/2023 - 4:06:45 AM",
                "kind": "text"
              }
            ]
          },
          {
            "name": "async"
          },
          {
            "name": "param",
            "text": [
              {
                "text": "",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "1} this is the page number of the return pending proposals",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "10} this is the limit of the return type of the pending proposals",
                "kind": "text"
              }
            ]
          },
          {
            "name": "returns",
            "text": [
              {
                "text": ">}",
                "kind": "text"
              }
            ]
          }
        ],
        "x-nestia-method": "GET"
      }
    },
    "/portals/{id}": {
      "get": {
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "description": "",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalWithBalance"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "x-nestia-namespace": "portals.getPortal",
        "x-nestia-jsDocTags": [],
        "x-nestia-method": "GET"
      }
    },
    "/portals/proposals": {
      "get": {
        "tags": [],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "description": "",
            "required": true
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "description": "",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Readonly__type.o1"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "summary": "Get all Pending proposals",
        "description": "The code snippet you provided is a method in the `PortalsController` class. It is a route handler\nfor the GET request to `/proposals` endpoint. Here's a breakdown of what it does:\nGets page",
        "security": [
          {
            "bearer": []
          }
        ],
        "x-nestia-namespace": "portals.proposals.getPendingProposals",
        "x-nestia-jsDocTags": [
          {
            "name": "summary",
            "text": [
              {
                "text": "Get all Pending proposals",
                "kind": "text"
              }
            ]
          },
          {
            "name": "date",
            "text": [
              {
                "text": "9/25/2023 - 4:06:45 AM",
                "kind": "text"
              }
            ]
          },
          {
            "name": "security",
            "text": [
              {
                "text": "bearer",
                "kind": "text"
              }
            ]
          },
          {
            "name": "async"
          },
          {
            "name": "param",
            "text": [
              {
                "text": "",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "1} this is the page number of the return pending proposals",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "10} this is the limit of the return type of the pending proposals",
                "kind": "text"
              }
            ]
          },
          {
            "name": "returns"
          }
        ],
        "x-nestia-method": "GET"
      },
      "post": {
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePortalProposalDto"
              }
            }
          },
          "required": true,
          "x-nestia-encrypted": false
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalProposals"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "summary": "Create a new proposal using user auth token to know which user is calling this function and sends email to user",
        "description": "The code snippet you provided is a method in the `PortalsController` class. It is a route handler\nfor the POST request to `/proposals` endpoint. Here's a breakdown of what it does:",
        "security": [
          {
            "bearer": []
          }
        ],
        "x-nestia-namespace": "portals.proposals.create",
        "x-nestia-jsDocTags": [
          {
            "name": "summary",
            "text": [
              {
                "text": "Create a new proposal using user auth token to know which user is calling this function and sends email to user",
                "kind": "text"
              }
            ]
          },
          {
            "name": "date",
            "text": [
              {
                "text": "9/25/2023 - 4:44:05 AM",
                "kind": "text"
              }
            ]
          },
          {
            "name": "async"
          },
          {
            "name": "param",
            "text": [
              {
                "text": "createPortalProposalDto",
                "kind": "text"
              }
            ]
          },
          {
            "name": "security",
            "text": [
              {
                "text": "bearer",
                "kind": "text"
              }
            ]
          },
          {
            "name": "returns"
          }
        ],
        "x-nestia-method": "POST"
      }
    },
    "/portals/proposals/{id}": {
      "get": {
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "description": "",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalProposals"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "x-nestia-namespace": "portals.proposals.getProposalById",
        "x-nestia-jsDocTags": [],
        "x-nestia-method": "GET"
      },
      "put": {
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "description": "",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePortalPropsalDto"
              }
            }
          },
          "required": true,
          "x-nestia-encrypted": false
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Http200Response"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "description": "it updates the proposal",
        "security": [
          {
            "bearer": []
          }
        ],
        "x-nestia-namespace": "portals.proposals.updateProposal",
        "x-nestia-jsDocTags": [
          {
            "name": "date",
            "text": [
              {
                "text": "9/25/2023 - 5:35:35 AM",
                "kind": "text"
              }
            ]
          },
          {
            "name": "security",
            "text": [
              {
                "text": "bearer",
                "kind": "text"
              }
            ]
          },
          {
            "name": "async"
          },
          {
            "name": "param",
            "text": [
              {
                "text": "id",
                "kind": "text"
              }
            ]
          },
          {
            "name": "returns"
          }
        ],
        "x-nestia-method": "PUT"
      }
    },
    "/portals/proposals/accept": {
      "get": {
        "tags": [],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "description": "",
            "required": true
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "description": "",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Readonly__type.o2"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "summary": "Retrieve a list of accepted Portal proposals\ndescription: Retrieve a list of accepted Portal proposals. The list supports pagination.\nparameters",
        "description": "The code snippet you provided is a method in the `PortalsController` class. It is a route handler\nfor the GET request to `/proposals/accept` endpoint. Here's a breakdown of what it does:\nGets page",
        "security": [
          {
            "bearer": []
          }
        ],
        "x-nestia-namespace": "portals.proposals.accept.getAcceptedProposals",
        "x-nestia-jsDocTags": [
          {
            "name": "summary",
            "text": [
              {
                "text": "Retrieve a list of accepted Portal proposals\ndescription: Retrieve a list of accepted Portal proposals. The list supports pagination.\nparameters",
                "kind": "text"
              }
            ]
          },
          {
            "name": "date",
            "text": [
              {
                "text": "9/25/2023 - 4:06:45 AM",
                "kind": "text"
              }
            ]
          },
          {
            "name": "security",
            "text": [
              {
                "text": "bearer",
                "kind": "text"
              }
            ]
          },
          {
            "name": "async"
          },
          {
            "name": "param",
            "text": [
              {
                "text": "",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "1} this is the page number of the return pending proposals",
                "kind": "text"
              }
            ]
          },
          {
            "name": "param",
            "text": [
              {
                "text": "",
                "kind": "parameterName"
              },
              {
                "text": " ",
                "kind": "space"
              },
              {
                "text": "10} this is the limit of the return type of the pending proposals",
                "kind": "text"
              }
            ]
          },
          {
            "name": "returns"
          }
        ],
        "x-nestia-method": "GET"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePortalDto": {
        "type": "object",
        "properties": {
          "address": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string"
          },
          "proposal_id": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string"
          }
        },
        "nullable": false,
        "required": [
          "address",
          "proposal_id"
        ],
        "x-typia-jsDocTags": []
      },
...
      "Readonly__type": {
        "type": "object",
        "properties": {
          "data": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortalWithBalance"
            }
          },
          "hasNextPage": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "boolean"
          }
        },
        "nullable": false,
        "required": [
          "data",
          "hasNextPage"
        ],
        "description": "Make all properties in T readonly",
        "x-typia-jsDocTags": []
      },
...
      "Readonly__type.o1": {
        "type": "object",
        "properties": {
          "data": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PortalProposals"
            }
          },
          "hasNextPage": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "boolean"
          }
        },
        "nullable": false,
        "required": [
          "data",
          "hasNextPage"
        ],
        "description": "Make all properties in T readonly",
        "x-typia-jsDocTags": []
      },
...
      "Http200Response": {
        "type": "object",
        "properties": {
          "message": {
            "x-typia-required": true,
            "x-typia-optional": false,
            "type": "string"
          }
        },
        "nullable": false,
        "required": [
          "message"
        ],
        "x-typia-jsDocTags": []
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}